pyproject.toml
[tool.poetry]
name = "chebs"
version = "0.1.0"
description = ""
authors = ["melody < 35177464+HexCodeFFF@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-chess = "^1.999"
ipython = "^8.3.0"
pyperclip = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
